split firmware into fetch and flash modules
